add utilities function

+++ Library/Utility/BitCount.cpp
@@ -1,5 +1,8 @@
-long long bitConut(long long n) {
-	long long c = 0;
+/**
+ *	立っているビットの数を数える
+ */
+ll bitConut(ll n) {
+	ll c = 0;
 	for (; n != 0; n &= n - 1) { ++c; }
 	return c;
 }
